+++ cpp/src/copying/gather.cu
@@ -7,7 +7,6 @@
 #include <cudf/table/table_view.hpp>
 #include <cudf/types.hpp>
 #include <cudf/utilities/type_dispatcher.hpp>
-#include <utilities/legacy/error_utils.hpp>
 
 #include <rmm/thrust_rmm_allocator.h>
 #include <thrust/count.h>
@@ -15,7 +14,6 @@
 #include <memory>
 
 namespace cudf {
-namespace experimental {
 namespace detail {
 struct dispatch_map_type {
   template <typename map_type,
@@ -92,17 +90,16 @@ std::unique_ptr<table> gather(table_view const& source_table,
 {
   CUDF_EXPECTS(gather_map.has_nulls() == false, "gather_map contains nulls");
 
-  std::unique_ptr<table> destination_table =
-    cudf::experimental::type_dispatcher(gather_map.type(),
-                                        dispatch_map_type{},
-                                        source_table,
-                                        gather_map,
-                                        gather_map.size(),
-                                        check_bounds,
-                                        ignore_out_of_bounds,
-                                        allow_negative_indices,
-                                        mr,
-                                        stream);
+  std::unique_ptr<table> destination_table = cudf::type_dispatcher(gather_map.type(),
+                                                                   dispatch_map_type{},
+                                                                   source_table,
+                                                                   gather_map,
+                                                                   gather_map.size(),
+                                                                   check_bounds,
+                                                                   ignore_out_of_bounds,
+                                                                   allow_negative_indices,
+                                                                   mr,
+                                                                   stream);
 
   return destination_table;
 }
@@ -118,5 +115,4 @@ std::unique_ptr<table> gather(table_view const& source_table,
   return detail::gather(source_table, gather_map, check_bounds, false, true, mr);
 }
 
-}  // namespace experimental
 }  // namespace cudf