+++ cpp/src/jit/launcher.h
@@ -17,10 +17,8 @@
  * limitations under the License.
  */
 
-#ifndef GDF_JIT_LAUNCHER_H
-#define GDF_JIT_LAUNCHER_H
+#pragma once
 
-#include <cudf/types.h>
 #include <jit/cache.h>
 #include <chrono>
 #include <fstream>
@@ -89,7 +87,6 @@ class launcher {
    *  contained in the members of `this`
    *
    * @tparam All parameters to launch the kernel
-   * @return Return GDF_SUCCESS if successful
    */
   template <typename... Args>
   void launch(Args... args)
@@ -108,5 +105,3 @@ class launcher {
 
 }  // namespace jit
 }  // namespace cudf
-
-#endif