+++ cpp/tests/utilities/base_fixture.cpp
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include <cudf/cudf.h>
+#include <cudf/types.hpp>
 
 namespace cudf {
 namespace test {